+++ src/compiler/frontmatter.mli
@@ -1,31 +1,49 @@
-type resolved = [ `Resolved ]
-type unresolved = [ `Unresolved ]
+type 'a loced := 'a * Cmarkit.Textloc.t
 
-type 'a fm = {
-  toplevel_attributes : Cmarkit.Attributes.t option;
-  math_link : 'a option;
-  theme : [ `Builtin of Themes.t | `External of string ] option;
-  css_links : 'a list;
-  js_links : 'a list;
-  dimension : (int * int) option;
-  highlightjs_theme : string option;
-  math_mode : [ `Mathjax | `Katex ] option;
-  external_ids : string list;
-}
+module Local : sig
+  type t = { toplevel_attributes : Cmarkit.Attributes.t Cmarkit.node option }
+  type 'a with_ = { x : 'a; fm : t }
 
-(** We use this trick to only allow [string fm] and [Asset.t fm], but it is
-    completely unnecessary and a flagrant example of useless over-engineering.
-*)
-type 'a t =
-  | Unresolved : string fm -> unresolved t
-  | Resolved : Asset.t fm -> resolved t
+  val empty : t
+  val with_empty : 'a -> 'a with_
+end
 
-module type Field = sig
+module Global : sig
+  type t = {
+    math_link : Asset.t loced option;
+    theme : [ `Builtin of Themes.t | `External of string ] loced option;
+    dimension : (int * int) loced option;
+    highlightjs_theme : string loced option;
+    math_mode : [ `Mathjax | `Katex ] loced option;
+    css_links : Asset.t list;
+    js_links : Asset.t list;
+    external_ids : string list;
+  }
+
+  type 'a with_ = { x : 'a; fm : t }
+
+  val empty : t
+  val with_empty : 'a -> 'a with_
+  val combine : t -> t -> t
+end
+
+type t = { local : Local.t; global : Global.t }
+
+val empty : t
+
+type fm := t
+
+module type Field := sig
   type t
 
   val key : string
-  val of_string : string * Cmarkit.Textloc.t -> (t, [ `Msg of string ]) result
-  val update_frontmatter : string fm -> t -> string fm
+
+  val of_string :
+    to_asset:(string -> Asset.t) ->
+    string * Cmarkit.Textloc.t ->
+    (t, [ `Msg of string ]) result
+
+  val update_frontmatter : fm -> t -> fm
 end
 
 module type Field_with_default := sig
@@ -35,21 +53,27 @@ module type Field_with_default := sig
 end
 
 module Toplevel_attributes :
-  Field_with_default with type t = Cmarkit.Attributes.t
+  Field_with_default with type t = Cmarkit.Attributes.t Cmarkit.node
 
-module Math_link : Field with type t = string
+module Math_link : Field with type t = Asset.t loced
 
 module Theme :
-  Field_with_default with type t = [ `Builtin of Themes.t | `External of string ]
+  Field_with_default
+    with type t = [ `Builtin of Themes.t | `External of string ] loced
 
-module Css_links : Field with type t = string list
-module Js_links : Field with type t = string list
-module Dimension : Field_with_default with type t = int * int
-module Hljs_theme : Field_with_default with type t = string
-module Math_mode : Field_with_default with type t = [ `Mathjax | `Katex ]
+module Css_links : Field with type t = Asset.t list
+module Js_links : Field with type t = Asset.t list
 
-val empty : resolved t
-val of_string : string -> int -> string -> unresolved t
+module Dimension : sig
+  include Field_with_default with type t = (int * int) loced
+
+  val of_string' : string * Cmarkit.Textloc.t -> (t, [ `Msg of string ]) result
+end
+
+module Hljs_theme : Field_with_default with type t = string loced
+module Math_mode : Field_with_default with type t = [ `Mathjax | `Katex ] loced
+
+val of_string : to_asset:(string -> Asset.t) -> string -> int -> string -> t
 
 type extraction = {
   frontmatter : string;
@@ -61,6 +85,3 @@ type extraction = {
 val extract : string -> extraction option
 (** Split the frontmatter and the rest of the input string, still computing
     offsets *)
-
-val combine : resolved t -> resolved t -> resolved t
-val resolve : unresolved t -> to_asset:(string -> Asset.t) -> resolved t