+++ src/compiler/slipshow.mli
@@ -8,6 +8,9 @@ module Action_plan = Action_plan
 type starting_state = int
 type delayed
 
+(** The additional string in the two functions below correspond to the warnings
+*)
+
 val delayed_to_string : delayed * string -> string
 val string_to_delayed : string -> (delayed * string) option
 
@@ -22,8 +25,8 @@ type file_reader = Fpath.t -> (string option, [ `Msg of string ]) result
       [Ok None]). *)
 
 val delayed :
+  ?options:Frontmatter.Global.t ->
   ?slipshow_js:Asset.t ->
-  ?frontmatter:Frontmatter.resolved Frontmatter.t ->
   ?file:string ->
   ?read_file:file_reader ->
   has_speaker_view:bool ->
@@ -37,10 +40,10 @@ val add_starting_state :
   ?autofocus:bool -> delayed -> starting_state option -> string
 
 val convert :
+  ?options:Frontmatter.Global.t ->
   has_speaker_view:bool ->
   ?autofocus:bool ->
   ?slipshow_js:Asset.t ->
-  ?frontmatter:Frontmatter.resolved Frontmatter.t ->
   ?file:string ->
   ?starting_state:starting_state ->
   ?read_file:file_reader ->